requirements.txt
python-docx>=1.1.0
reportlab>=4.0.0
openai>=1.0.0
python-dotenv>=1.0.0